requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
tdqm==0.0.1
tqdm==4.66.1
urllib3==2.0.5
